skip signal incase of pytest run

+++ ebookapp/signals.py
@@ -2,6 +2,8 @@ import boto3
 from django.dispatch import receiver
 from django.db.models.signals import post_save,post_delete
 from .models import Book
+from django.conf import settings
+import os
 
 
 
@@ -14,6 +16,9 @@ def sync_book_stock_to_dynamodb(sender, instance, **kwargs):
   function will recieve the signal when post_save called on Mysql
   will sync the dynamodb on save operation
   """
+    # Skip AWS calls during tests
+  if settings.DEBUG or os.environ.get("PYTEST_CURRENT_TEST"):
+      return
   dynamodb = boto3.resource('dynamodb')
   book_table = dynamodb.Table('Books_Stock')
   book_table.put_item(
@@ -33,6 +38,8 @@ def delete_books_stock_from_dynamodb(sender,instance,**kwargs):
   function will recieve the delete signal on any book entry delete
   will remove the stock entry from dynamodb
   """
+  if settings.DEBUG or os.environ.get("PYTEST_CURRENT_TEST"):
+    return
   dynamodb = boto3.resource('dynamodb')
   book_table = dynamodb.Table('Books_Stock')
   book_table.delete_item(
